split manifests into areas
