+++ src/packages/settings/advanced/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_ADVANCED_SETTINGS_MENU_ALIAS = 'Umb.Menu.AdvancedSettings';
