+++ src/packages/settings/section/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_SETTINGS_SECTION_ALIAS = 'Umb.Section.Settings';
