+++ src/packages/settings/structure/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_STRUCTURE_SETTINGS_MENU_ALIAS = 'Umb.Menu.StructureSettings';
