+++ src/packages/settings/umbraco-package.ts
@@ -1,2 +1,9 @@
 export const name = 'Umbraco.Core.Settings';
-export const extensions = [];
+export const extensions = [
+	{
+		name: 'Settings Bundle',
+		alias: 'Umb.Bundle.Settings',
+		type: 'bundle',
+		js: () => import('./manifests.js'),
+	},
+];
